Make Lexilla c++11 compatible

+++ lexers/LexCPP.cxx
@@ -529,7 +529,7 @@ const LexicalClass lexicalClasses[] = {
 	27, "SCE_C_ESCAPESEQUENCE", "literal string escapesequence", "Escape sequence",
 };
 
-constexpr int sizeLexicalClasses{ Sci::size(lexicalClasses) };
+constexpr int sizeLexicalClasses{ static_cast<int>(Sci::size(lexicalClasses)) };
 
 }
